RT UDP packet RX, timer sync, plot pan/zoom, RESET
Real time UDP packets receiving and displaying (datagrams are not stored in the GUI to save memory, make sure to do so on the transmitter end if you wish, e.g. use a RTDB);
Timer is synched to the UDP datagram readings and updated every second in case there are no datagrams to be received;
Graphs now have a button to the top-right that enables/disables panning and zooming mode;
RESET button implemented to clear the graphs and values displayed and restart the timer;
START button removed due to real time functionalities being implemented, now the program is constantly looking to receive and display data.